Fix adding location and category to tool page
Can now select the location and category. Stores the location id and the
category id.

Really helpful site on understanding relations in Bookshelf. Used in
models to set them up properly.
http://stackabuse.com/bookshelf-js-a-node-js-orm/

controllers/tools.js = Helpful site for using withRelated and fetchAll
(you can't, so use a Collection)
http://stackoverflow.com/questions/35441158/bookshelf-js-fetchall-withrelated-option-returns-empty-relation-object

helpers/getCategories.js = switched to an array of objects and include
id and name

helpers/getLocations.js = switched to an array of objects and include
id and name. Commented code is to change it to just an object, which may
be useful later.

models/categories.js = Added the foreign key, 'category_id', and changed
so it works with relations

models/locations.js = Added the foreign keys, 'locations_id', and changed
so it works with relations

models/people.js = Added the foreign key, 'locations_id', and changed
so it works with relations

models/tools.js = Added the foreign keys, 'category_id' and
'locations_id', and changed so it works with relations

views/tool.pug = fixed the category and location select each statements
to check on the category_id value from the Tools table against the
id value from the Category table. Likewise for locations. Also changed
the column settings (MaterializeCSS) so the save and delete buttons line
up better in smaller windows.