Add error handling for deleting dependant item
When a tool is assigned to a location or category, then that location or
category can not be deleted because of foreign keys.

This quick fix just shows an error message with a link to the home page.
Better option would be to have the user select a new location or
category before allowing the delete.

Closes #6

Files Changed:

controllers/categories.js & controllers/locations.js =
        - add a catch to the delete method which renders the error page
        - change catches in edit and listAll methods to render the error
          page instead of console.log

views/error.pug =
        - add style and error messages