Adds new fields to people record
Adds the following fields: office hours, office address, bio and
website.

Also changes the field title to moniker (so that it doesn't mess with
the title variable passed in from the controller that controls the title
of the page and the background color of the top nav section).

Creates a new schema.sql file.

Adds a jQuery call in the footer to refresh the bio textarea so that it
auto-expands the textarea.

Adds a style to the avatar image on the person page to make the image
width 100% so it fills up the column it is in.

Adds additional fields in People table, remname title to moniker.