Add reset password and unique username/email functionality
Adds a link to the login page for reseting password. Clicking the link takes
the user to a page where they type in their username or email, then clicks a
reset button.  This then resets the password in the database and sends an email
to address for the user. This requires the email and username to be unique, so
altered the database to make that so.

- routes/auth.js
  - add check for unique email when registering a new user
  - add routes for the reset password functionality.
- views/login.pug
  - show an error or message if the username/password is wrong
  - adds a link to reset the password
- views/register.pug
  - Add message if email is already taken
- views/user.pug
  - update logic to account for trying to register a new user with an existing username or email
- veiws/reset.pug
  - the reset password form
- package.json, package-lock.json
  - add the nodemailer package
- server.js
  - add the nodemailer package
- helpers/mailer.js
  - initialize the mailer method
  - pulls email account from the .env file
  - NOTE: set the email account in the .env file
- .knex/migrations/20170831140041_user_email_unique.js
  - makes the Users.email a unique non-null item.