Add links to the API calls and change icons
If logged in as an admin, this shows links to the api calls for
locations, tools, people, and categories-locations JSON files. This
makes it so much easier to check how the site returns the JSON API
calls, rather than typing in the URL.

Also changes the icon for locations to location_city because it looks
more like a building than the building icon.